Made water level a float

+++ engine/src/main/battlecode/world/GameWorld.java
@@ -31,7 +31,7 @@ public strictfp class GameWorld {
     private int[] pollution;
     private int[] dirt;
     private int initialWaterLevel;
-    private int waterLevel;
+    private float waterLevel;
     private boolean[] flooded;
     private InternalRobot[][] robots;
     private final LiveMap gameMap;
@@ -381,7 +381,7 @@ public strictfp class GameWorld {
      * Updates the global water level according to an arbitrary function.
      */
     public void updateWaterLevel() {
-        this.waterLevel = (int) Math.floor(Math.pow(this.currentRound / 200.0, 2));
+        this.waterLevel = (float) Math.floor(Math.pow(this.currentRound / 200.0, 2));
     }
 
     // ***********************************
@@ -573,8 +573,8 @@ public strictfp class GameWorld {
                                 setWinnerArbitrary();
 
         // update the round statistics
-         matchMaker.addTeamSoup(Team.A, teamInfo.getSoup(Team.A));
-         matchMaker.addTeamSoup(Team.B, teamInfo.getSoup(Team.B));
+        matchMaker.addTeamSoup(Team.A, teamInfo.getSoup(Team.A));
+        matchMaker.addTeamSoup(Team.B, teamInfo.getSoup(Team.B));
 
         if (gameStats.getWinner() != null)
             running = false;
